+++ test/vhpi/vhpi_test.h
@@ -58,6 +58,7 @@ void vhpi15_startup(void);
 void vhpi16_startup(void);
 void vhpi17_startup(void);
 void vhpi18_startup(void);
+void vhpi19_startup(void);
 void issue744_startup(void);
 void issue762_startup(void);
 void issue978_startup(void);