+++ NEWS.md
@@ -6,6 +6,7 @@
 - VHPI value change callbacks for indexed names now work correctly (#1428).
 - Added support for iterating physical unit declarations in VHPI (#1437).
 - Several other minor bugs were resolved (#1431, #1438).
+- VHPI returns correct format for time and time_vector types (#1441).
 
 ## Version 1.19.2 - 2026-02-22
 - Fixed a crash when dumping a waveform `--dump-arrays` and the design