+++ NEWS.md
@@ -1,5 +1,6 @@
 ## Unreleased changes
 - VHPI value change callbacks for indexed names now work correctly (#1428).
+- Added support for iterating physical unit declarations in VHPI (#1437).
 - Several other minor bugs were resolved (#1431, #1438).
 
 ## Version 1.19.2 - 2026-02-22